third_party/blink/web_tests/wpt_internal/bluetooth/requestDevice/name-empty-device-from-name-prefix-filter.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
bluetooth_test(t => {
  return setBluetoothFakeAdapter('EmptyNameHeartRateAdapter')
      .then(
          () => promise_rejects_dom(
              t, 'NotFoundError',
              requestDeviceWithTrustedClick(
                  {filters: [{namePrefix: 'a', services: ['heart_rate']}]})));
}, 'An empty name device is not matched by a filter with a namePrefix.');
</script>
